gsdx-ogl: use 64 bits counter + fix division factor
I also added a counter of the real size of the texture.

I have a bad overhead for pbo transfer

+++ plugins/GSdx/GSUniformBufferOGL.h
@@ -24,7 +24,7 @@
 #include "GLState.h"
 
 #ifdef ENABLE_OGL_DEBUG_MEM_BW
-extern uint32 g_uniform_upload_byte;
+extern uint64 g_uniform_upload_byte;
 #endif
