+++ plugins/GSdx/GSVertexArrayOGL.h
@@ -24,7 +24,7 @@
 #include "config.h"
 
 #ifdef ENABLE_OGL_DEBUG_MEM_BW
-extern uint32 g_vertex_upload_byte;
+extern uint64 g_vertex_upload_byte;
 #endif
 
 struct GSInputLayoutOGL {
